chore: update the actual multi_clone script with the updated setup
